base bloc for feed created

+++ lib/features/presentation/cubits/den/comments/comment_bloc.dart
@@ -11,11 +11,11 @@ class CommentBloc extends Bloc<CommentEvent, CommentState> {
   CommentBloc(this.repository) : super(const CommentState()) {
     on<FetchComments>(_onFetchComments);
     on<AddComment>(_onAddComment);
-     on<DeleteComment>(_onDeleteComment);
+    on<DeleteComment>(_onDeleteComment);
     on<AddCommentSuccess>(_onAddCommentSuccess);
     on<AddCommentFailure>(_onAddCommentFailure);
-     on<DeleteCommentSuccess>(_onDeleteCommentSuccess);
-      on<DeleteCommentFailure>(_onDeleteCommentFailure);
+    on<DeleteCommentSuccess>(_onDeleteCommentSuccess);
+    on<DeleteCommentFailure>(_onDeleteCommentFailure);
   }
 
   Future<void> _onFetchComments(
@@ -28,6 +28,7 @@ class CommentBloc extends Bloc<CommentEvent, CommentState> {
 
       log("Comments for user post ${comments.length}");
       emit(state.copyWith(
+        postId: event.postId,
         comments: comments,
         isLoadingComments: false,
       ));
@@ -45,28 +46,26 @@ class CommentBloc extends Bloc<CommentEvent, CommentState> {
     AddComment event,
     Emitter<CommentState> emit,
   ) async {
-    emit(state.copyWith(isPostingComment: true, error: null));
+    emit(state.copyWith(postId: event.comment.postId, isPostingComment: true, error: null));
 
     // Immediately update UI locally
     final updatedComments = [...state.comments, event.comment];
-    
+
     emit(state.copyWith(comments: updatedComments));
 
     try {
       await repository.addComment(comment: event.comment);
-      add(AddCommentSuccess(event.comment));
+      add(AddCommentSuccess(postId:  event.comment.postId!, comment:  event.comment));
     } catch (e) {
       add(AddCommentFailure(e.toString()));
     }
   }
 
- 
-
   void _onAddCommentSuccess(
     AddCommentSuccess event,
     Emitter<CommentState> emit,
   ) {
-    emit(state.copyWith(isPostingComment: false));
+    emit(state.copyWith(postId: event.comment.postId, isPostingComment: false));
   }
 
   void _onAddCommentFailure(
@@ -76,36 +75,38 @@ class CommentBloc extends Bloc<CommentEvent, CommentState> {
     emit(state.copyWith(isPostingComment: false, error: event.error));
   }
 
-
 // Delete comment
   Future<void> _onDeleteComment(
-  DeleteComment event,
-  Emitter<CommentState> emit,
-) async {
-
-  // Optimistically remove from UI
-  final updatedComments = state.comments.where((c) => c.id != event.commentId).toList();
-  emit(state.copyWith(comments: updatedComments));
-
-  try {
-    await repository.deleteComment(commentID: event.commentId);
-    add(DeleteCommentSuccess(event.commentId));
-  } catch (e) {
-    add(DeleteCommentFailure(e.toString()));
+    DeleteComment event,
+    Emitter<CommentState> emit,
+  ) async {
+    try {
+      print("delete comment from id ${event.commentId}");
+      bool successDelete =
+          await repository.deleteComment(commentID: event.commentId);
+      if (successDelete) {
+        // Optimistically remove from UI
+        final updatedComments =
+            state.comments.where((c) => c.id != event.commentId).toList();
+        emit(state.copyWith(postId: event.commentId, comments: updatedComments));
+        add(DeleteCommentSuccess(event.commentId));
+      }
+    } catch (e) {
+      add(DeleteCommentFailure(e.toString()));
+    }
   }
-}
 
-void _onDeleteCommentSuccess(
-  DeleteCommentSuccess event,
-  Emitter<CommentState> emit,
-) {
-  emit(state.copyWith(isDeletingComment: false));
-}
+  void _onDeleteCommentSuccess(
+    DeleteCommentSuccess event,
+    Emitter<CommentState> emit,
+  ) {
+    emit(state.copyWith(isDeletingComment: false));
+  }
 
-void _onDeleteCommentFailure(
-  DeleteCommentFailure event,
-  Emitter<CommentState> emit,
-) {
-  emit(state.copyWith(isDeletingComment: false, error: event.error));
-}
+  void _onDeleteCommentFailure(
+    DeleteCommentFailure event,
+    Emitter<CommentState> emit,
+  ) {
+    emit(state.copyWith(isDeletingComment: false, error: event.error));
+  }
 }
